+++ src/ts.rs
@@ -1,0 +1,15 @@
+use lsp_types::{Position, Range};
+use tree_sitter::Node;
+
+pub fn get_range_from_node(node: Node) -> Range {
+    Range {
+        start: Position {
+            line: node.start_position().row as u32,
+            character: node.start_position().column as u32,
+        },
+        end: Position {
+            line: node.end_position().row as u32,
+            character: node.end_position().column as u32,
+        },
+    }
+}
